reorganizing based on reviewer feedback
