+++ cmd/auth/auth.go
@@ -16,7 +16,7 @@ import (
 
 	"github.com/charmbracelet/log"
 	"github.com/datarobot/cli/internal/assets"
-	"github.com/datarobot/cli/internal/baseauth"
+	"github.com/datarobot/cli/internal/config"
 	"github.com/spf13/viper"
 )
 
@@ -109,7 +109,7 @@ func writeConfigFile() {
 
 func GetAPIKey() (string, error) {
 	// Returns the API key if there is one, otherwise returns an empty string
-	key := viper.GetString(baseauth.DataRobotAPIKey)
+	key := viper.GetString(config.DataRobotAPIKey)
 
 	return key, nil
 }
