+++ cmd/auth/logout.go
@@ -10,13 +10,13 @@ package auth
 
 import (
 	"github.com/charmbracelet/log"
-	"github.com/datarobot/cli/internal/baseauth"
+	"github.com/datarobot/cli/internal/config"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 )
 
 func LogoutAction() error {
-	viper.Set(baseauth.DataRobotAPIKey, baseauth.DataRobotAPIKey)
+	viper.Set(config.DataRobotAPIKey, config.DataRobotAPIKey)
 
 	writeConfigFile()
